\name{pep_hxb2JPT}
\alias{pep_hxb2JPT}

\title{
	Peptide dataset for hxb2 and JPT clades
}

\description{
A peptide dataset created from gpr files .

\describe{
\item{peptides:}{The rownames are 15mers from SIV proteins}
\item{space:}{An information regarding the genomic location of the peptide.}
\item{ranges:}{The start and end coordinate of the peptide based on the coordinates of the reference parent protein.}
\item{zSums:}{The sum of the zScale of each amino-acid of the 15mer.}
}
}
\note{
	The excel sheet used to create this collection is available in the extdata/raw folder.
}

\keyword{datasets}